WindowsAppSDK-ProjectTemplates/python/packaged/pyproject.toml
[project]
name = "packaged-winapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "winappsdk-Foundation",
    "winappsdk-InteractiveExperiences"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
find-links = ["../wheels"]